+++ tests/setter_behavior_blinker_test.cpp
@@ -146,7 +146,7 @@ TEST(SetterBehaviorBlinkerTest, ShouldTruncateToNumDigits)
 
 TEST(SetterBehaviorBlinkerTest, ShouldRejectTooHighNumDigits)
 {
-    const NumDigits       max_digits = max_digits_for_base(NumberBase::DEC);
+    const NumDigits       max_digits = max_digits_for_radix(Radix::DEC);
     const TestBlinkerCase test_case{
         .trace_check = TraceCheck::ENFORCED,
         .numDigits   = max_digits + 1,